Renaming package from update_available to update.
