+++ ui-lib/build.gradle.kts
@@ -46,7 +46,7 @@ android {
                     "src/main/res/ui/send",
                     "src/main/res/ui/settings",
                     "src/main/res/ui/support",
-                    "src/main/res/ui/update_available",
+                    "src/main/res/ui/update",
                     "src/main/res/ui/wallet_address"
                 )
             )
