+++ ui-lib/src/androidTest/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:name=".screen.backup.TestBackupActivity"
             android:exported="false" />
         <activity
-            android:name=".screen.update_available.TestUpdateAvailableActivity"
+            android:name=".screen.update.TestUpdateAvailableActivity"
             android:exported="false" />
     </application>
