+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/update/RestoreTag.kt
@@ -1,0 +1,10 @@
+package co.electriccoin.zcash.ui.screen.update
+
+/**
+ * These are only used for automated testing.
+ */
+object RestoreTag {
+    const val BTN_LATER = "btn_later"
+    const val BTN_DOWNLOAD = "btn_download"
+    const val PROGRESSBAR_DOWNLOADING = "progressbar_downloading"
+}
