+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/update/model/UpdateState.kt
@@ -1,0 +1,9 @@
+package co.electriccoin.zcash.ui.screen.update.model
+
+enum class UpdateState {
+    Prepared,
+    Running,
+    Failed,
+    Done,
+    Canceled
+}
