--- ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/update_available/RestoreTag.kt
@@ -1,11 +1,0 @@
-@file:Suppress("PackageNaming")
-package co.electriccoin.zcash.ui.screen.update_available
-
-/**
- * These are only used for automated testing.
- */
-object RestoreTag {
-    const val BTN_LATER = "btn_later"
-    const val BTN_DOWNLOAD = "btn_download"
-    const val PROGRESSBAR_DOWNLOADING = "progressbar_downloading"
-}
