--- ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/update_available/model/UpdateState.kt
@@ -1,10 +1,0 @@
-@file:Suppress("PackageNaming")
-package co.electriccoin.zcash.ui.screen.update_available.model
-
-enum class UpdateState {
-    Prepared,
-    Running,
-    Failed,
-    Done,
-    Canceled
-}
