+++ ui-lib/src/main/res/ui/update/drawable/update_main_graphic.xml
