--- ui-lib/src/main/res/ui/update_available/values/strings.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="update_available_header">Update available!</string>
-    <string name="update_available_critical_header">Critical update required!</string>
-    <string name="update_available_image_content_description"></string>
-    <string name="update_available_description">There is a new version of the app available.</string>
-    <string name="update_available_link_text">Learn more about this update here.</string>
-    <string name="update_available_download_button">Download Update</string>
-    <string name="update_available_later_enabled_button">Remind me later</string>
-    <string name="update_available_later_disabled_button">This can not be skipped.</string>
-</resources>
